requirements.txt
mcp
fastmcp
ollama
pydantic
duckduckgo-search
psutil
numpy
PyYAML
